// types/zod/outputTypeSchemas/businessFindManyArgsSchema.ts
export const businessSelectSchema: z.ZodType<Prisma.businessSelect> = z.object({
  business_id: z.boolean().optional(),
  address_id: z.boolean().optional(),
  delivery_address_id: z.boolean().optional(),
  finance_id: z.boolean().optional(),
  type: z.boolean().optional(),
  is_business_unit: z.boolean().optional(),
  business_group_name: z.boolean().optional(),
  name: z.boolean().optional(),
  description: z.boolean().optional(),
  tax_id: z.boolean().optional(),
  registration_id: z.boolean().optional(),
  email: z.boolean().optional(),
  telephone: z.boolean().optional(),
  telephone_code: z.boolean().optional(),
  telephone_number: z.boolean().optional(),
  website_url: z.boolean().optional(),
  working_hours: z.boolean().optional(),
  seats: z.boolean().optional(),
  minimum_order: z.boolean().optional(),
  offers_daily_meals: z.boolean().optional(),
  offers_daily_meals_on_weekend: z.boolean().optional(),
  popular: z.boolean().optional(),
  new: z.boolean().optional(),
  created_at: z.boolean().optional(),
  updated_at: z.boolean().optional(),
  parent_business_id: z.boolean().optional(),
  reviewable_id: z.boolean().optional(),
  stripe_account_id: z.boolean().optional(),
  stripe_customer_id: z.boolean().optional(),
  word_buy_stripe_product_id: z.boolean().optional(),
  word_buy_stripe_subscription_id: z.boolean().optional(),
  daily_users_sorted: z.boolean().optional(),
  daily_users_sorting_type: z.boolean().optional(),
  restaurant_overwhelmed: z.boolean().optional(),
  first_activated_at: z.boolean().optional(),
  active: z.boolean().optional(),
  sales_representative_id: z.boolean().optional(),
  fiscal_devices_id: z.boolean().optional(),
  purchase_order_limit_amount: z.boolean().optional(),
  address: z.union([z.boolean(),z.lazy(() => addressesArgsSchema)]).optional(),
  delivery_address: z.union([z.boolean(),z.lazy(() => addressesArgsSchema)]).optional(),
  finances: z.union([z.boolean(),z.lazy(() => financesArgsSchema)]).optional(),
  business_users: z.union([z.boolean(),z.lazy(() => business_usersFindManyArgsSchema)]).optional(),
  documents: z.union([z.boolean(),z.lazy(() => documentsFindManyArgsSchema)]).optional(),
  parent_business: z.union([z.boolean(),z.lazy(() => businessArgsSchema)]).optional(),
  child_businesses: z.union([z.boolean(),z.lazy(() => businessFindManyArgsSchema)]).optional(),
  taxi_orders: z.union([z.boolean(),z.lazy(() => taxi_ordersFindManyArgsSchema)]).optional(),
  delivery_orders: z.union([z.boolean(),z.lazy(() => delivery_ordersFindManyArgsSchema)]).optional(),
  menus: z.union([z.boolean(),z.lazy(() => menusFindManyArgsSchema)]).optional(),
  reviewable: z.union([z.boolean(),z.lazy(() => reviewableArgsSchema)]).optional(),
  word_buy_stripe_product: z.union([z.boolean(),z.lazy(() => local_productsArgsSchema)]).optional(),
  word_buys: z.union([z.boolean(),z.lazy(() => word_buyFindManyArgsSchema)]).optional(),
  reservations: z.union([z.boolean(),z.lazy(() => reservationsFindManyArgsSchema)]).optional(),
  promo_sections: z.union([z.boolean(),z.lazy(() => promo_sections_buyFindManyArgsSchema)]).optional(),
  business_teams: z.union([z.boolean(),z.lazy(() => business_teamsFindManyArgsSchema)]).optional(),
  business_order_lobbies: z.union([z.boolean(),z.lazy(() => order_lobbiesFindManyArgsSchema)]).optional(),
  user_favorite_businesses: z.union([z.boolean(),z.lazy(() => user_favorite_businessesFindManyArgsSchema)]).optional(),
  scoring_points: z.union([z.boolean(),z.lazy(() => scoring_pointsFindManyArgsSchema)]).optional(),
  daily_meal_drivers: z.union([z.boolean(),z.lazy(() => delivery_driversFindManyArgsSchema)]).optional(),
  late_events: z.union([z.boolean(),z.lazy(() => late_eventsFindManyArgsSchema)]).optional(),
  fiscal_device: z.union([z.boolean(),z.lazy(() => fiscal_devicesArgsSchema)]).optional(),
  daily_meals_subscribers: z.union([z.boolean(),z.lazy(() => daily_meals_subscriptionsFindManyArgsSchema)]).optional(),
  account_actions: z.union([z.boolean(),z.lazy(() => account_actionsFindManyArgsSchema)]).optional(),
  business_clients: z.union([z.boolean(),z.lazy(() => business_clientsFindManyArgsSchema)]).optional(),
  _count: z.union([z.boolean(),z.lazy(() => BusinessCountOutputTypeArgsSchema)]).optional(),
}).strict()

export const businessFindManyArgsSchema: z.ZodType<Prisma.businessFindManyArgs> = z.object({
  select: businessSelectSchema.optional(),
  include: businessIncludeSchema.optional(),
  where: businessWhereInputSchema.optional(),
  orderBy: z.union([ businessOrderByWithRelationInputSchema.array(),businessOrderByWithRelationInputSchema ]).optional(),
  cursor: businessWhereUniqueInputSchema.optional(),
  take: z.number().optional(),
  skip: z.number().optional(),
  distinct: z.union([ BusinessScalarFieldEnumSchema,BusinessScalarFieldEnumSchema.array() ]).optional(),
}).strict() ;